Refactored into seperate classes, and added network evaluation test. Evaluation is tested and working for a simple network with 1 input, 1 bias, 1 output, 1 hidden, and 4 connections. Also added NeuralNet::register_sigmoid so that different sigmoids can be easily registered.
