+++ tests/LoadNodeSensor.cc
@@ -1,22 +1,20 @@
 #include <gtest/gtest.h>
 #include "Genome.hh"
-#include "Network.hh"
-#include "Node.hh"
 
 TEST(Network,LoadInputs){
   //Network net(std::make_shared<Genome>(0));
 
-  std::vector<double> inputs = {1,2,3};
-  auto genome = make_genome(0);
-  for (auto i=0u; i<inputs.size(); i++) {
-    genome.AddNode(Node::Type::Sensor,Node::Function::Input);
-    //net.inputs.emplace_back(std::make_shared<Node>(Node::Type::Sensor,Node::Function::Input));
-  }
-  Network net(std::make_shared<Genome>(genome));
-  net.LoadInputs(inputs);
-  auto it = inputs.begin();
-
-  // for (auto& inode : net.inputs) {
-  //   ASSERT_EQ(inode->activation,*it++);
+  // std::vector<double> inputs = {1,2,3};
+  // auto genome = make_genome(0);
+  // for (auto i=0u; i<inputs.size(); i++) {
+  //   genome.AddNode(Node::Type::Sensor,Node::Function::Input);
+  //   //net.inputs.emplace_back(std::make_shared<Node>(Node::Type::Sensor,Node::Function::Input));
   // }
+  // Network net(std::make_shared<Genome>(genome));
+  // net.LoadInputs(inputs);
+  // auto it = inputs.begin();
+
+  // // for (auto& inode : net.inputs) {
+  // //   ASSERT_EQ(inode->activation,*it++);
+  // // }
 }
